don't truncate patdef at 7 bits
The biggest game so far to use Pently is Nova the Squirrel, which
is also the first game to use more than 128 patterns.  The patdef
macro was incorrectly truncating an offset into an array of pattern
data pointers to 8 bits before dividing it by 2, causing playPat*
commands to use patterns 0-126 instead of 128-254.
Credit to Nova the Squirrel programmer NovaSquirrel for the fix.