fix dubious rate_divider_cancel behavior
On frames when the sound effect doesn't transition from one step to
the next, pently_update_one_ch was skipping over most of the mixer.
This was fine in Pently 1 with no sound effects, but I have no idea
how it went undiscovered until now.  I imagine it'd cause incorrect
playback when a note and a sound effect with a "rate" (ticks per
step) greater than 1 are playing on the same channel.