optimize transpose runs in fallthrough group
In each fallthrough group (consecutive patterns that fall through
combined with the next pattern), find a set of patterns including
the last pattern that fit in a 2-octave range if possible, then
set the group's transposition base to the lowest note in that set.

This should fix #1, though I'll have to leave a comment about the
"pickup"-in-pattern issue.