pentlybss: add ASM6 mode
ASM6 mode adds dsb statements for pentlymusicbase, as well as all
statically allocated ZP and BSS variables, to be included within
an enum...ende block in the Application.

Caveat: Only BSS, not zero page, is dynamic this way.  Until
there's a way to have pentlyzp.inc see pentlyconfig.inc, ZP uses
the largest of the three layouts (32 bytes), as opposed to the
16-byte Pently 1 layout without support for music or the 21-byte
Pently 3 layout without support for attack envelopes.

for #9