recognize Hyper Click mouse
Unlike Nintendo's Super NES Mouse, Hyperkin's Hyper Click mouse
doesn't respond to sensitivity changes requested by the host.
So instead of attempting to detect the mouse by cycling its
sensitivity to medium, recognize any sensitivity with the mouse
signature bits 0001 as a mouse.

To compensate, display the current sensitivity in the footer and
allow changing it by clicking it.