add list of fields in rough access order
Processors descended from Intel 8080, such as the Z80 in the Sega
Master System or the SM83 in the Game Boy, are more efficient at
accessing data stored sequentially.  This means accessing fields of a
struct or properties is more efficient if the order in RAM of these
properties reflects the order in which they are read and written.
Document these fields and their order in case Pently gets ported to
such a platform.

Seeing as this is loosely related to pentlybss.py, I saw some
cosmetic improvements I could make to the latter.  And while testing
that those improvements made no change to the program, I identified
a file extension that make clean was neglecting to delete.