allow dotted duration in time sig denominator
In a common practice time signature, "3/4" means 3 beats per
measure and the quarter note gets one beat.  But in compound
prolation, the beat is a dotted note.  The time signature
indicates this by tripling the numerator and writing one-third
of a beat as the denominator: usually "6/8" or "9/8".
But perhaps an easier to understand way to express it, inspired
by the work of composer Carl Orff, is as "2/4." where the
denominator is drawn as a dotted quarter note.

Fix #44