docs: don't use "tick" for a row
"Tick" in trackers tends to mean one update of the sound driver.
In MIDI and sample-based tracker formats (mod, s3m, xm, it),
this is usually a 96th note. In chiptune tracker formats
(ftm and the like), it's one video field (16.4 to 20 ms).
Either way, it's much shorter than a "row" given that the
"speed" value is defined as the length of a row in ticks.